Updated project
Cleaned up comments
removed import re, I wasn't using it
Fixed bug in placeShips() that wasn't allowing ships to be placed all the way to the edges
placeShips now holds the number representation of the ship for each spot instead of just a '*'.  This allows for easier debugging.
Updated the rest of the project to check for spaces instead of stars in computerShips